Move Server related code into a new servers module
