+++ crates/core/tedge_actors/src/actors.rs
@@ -1,9 +1,4 @@
-use crate::message_boxes::MessageReceiver;
-use crate::ConcurrentServerMessageBox;
-use crate::Message;
 use crate::RuntimeError;
-use crate::Sender;
-use crate::ServerMessageBox;
 use async_trait::async_trait;
 
 /// Enable a struct to be used as an actor.
@@ -27,178 +22,6 @@ pub trait Actor: 'static + Send + Sync {
     /// updating internal state,
     /// and sending messages to peers.
     async fn run(&mut self) -> Result<(), RuntimeError>;
-}
-
-/// An actor that wraps a request-response server
-///
-/// Requests are processed in turn, leading either to a response or an error.
-pub struct ServerActor<S: Server> {
-    server: S,
-    messages: ServerMessageBox<S::Request, S::Response>,
-}
-
-impl<S: Server> ServerActor<S> {
-    pub fn new(server: S, messages: ServerMessageBox<S::Request, S::Response>) -> Self {
-        ServerActor { server, messages }
-    }
-}
-
-/// A server defines the behavior of an actor processing requests sending back responses.
-///
-/// A `Server` is defined by:
-/// - a `Request` message type for the requests received from client actors,
-/// - a `Response` message type for the responses sent back to the requesters,
-/// - an asynchronous `handle` method that defines how the server responds to a request,
-///   updating its state and possibly performing side effects.
-///
-/// ```
-/// # use crate::tedge_actors::{Server, SimpleMessageBox};
-/// # use async_trait::async_trait;
-///
-/// # use crate::tedge_actors::examples;
-/// # type Operation = examples::calculator::Operation;
-/// # type Update = examples::calculator::Update;
-///
-/// /// State of the calculator server
-/// #[derive(Default)]
-/// struct Calculator {
-///     state: i64,
-/// }
-///
-/// /// Implementation of the calculator behavior
-/// #[async_trait]
-/// impl Server for Calculator {
-///
-///     type Request = Operation;
-///     type Response = Update;
-///
-///     fn name(&self) -> &str {
-///         "Calculator"
-///     }
-///
-///     async fn handle(&mut self, request: Self::Request) -> Self::Response {
-///         // Act accordingly to the request
-///         let from = self.state;
-///         let to = match request {
-///            Operation::Add(x) => from + x,
-///            Operation::Multiply(x) => from * x,
-///         };
-///
-///         // Update the server state
-///         self.state = to;
-///
-///         // Return the response
-///         Update{from,to}
-///     }
-/// }
-/// ```
-///
-/// To be used as an actor, a `Server` is wrapped into a [ServerActor](crate::ServerActor)
-///
-/// ```
-/// # use tedge_actors::{Actor, Builder, NoConfig, MessageReceiver, Sender, ServerActor, SimpleMessageBox, SimpleMessageBoxBuilder};
-/// use tedge_actors::test_helpers::ServiceProviderExt;
-/// # use crate::tedge_actors::examples::calculator::*;
-/// #
-/// # #[tokio::main]
-/// # async fn main_test() {
-/// #
-/// // As for any actor, one needs a bidirectional channel to the message box of the server.
-/// let mut actor_box_builder = SimpleMessageBoxBuilder::new("Actor", 10);
-/// let mut client_box = actor_box_builder.new_client_box(NoConfig);
-/// let server_box = actor_box_builder.build();
-///
-/// // Create an actor to handle the requests to a server
-/// let mut calculator_box = SimpleMessageBoxBuilder::new("Calculator - REMOVE ME", 16).build();
-/// let server = Calculator::new(calculator_box);
-/// let mut actor = ServerActor::new(server, server_box);
-///
-/// // The actor is then spawn in the background with its message box.
-/// tokio::spawn(async move { actor.run().await } );
-///
-/// // One can then interact with the actor
-/// // Note that now each request is prefixed by a number: the id of the requester
-/// client_box.send((1,Operation::Add(4))).await.expect("message sent");
-/// client_box.send((2,Operation::Multiply(10))).await.expect("message sent");
-/// client_box.send((1,Operation::Add(2))).await.expect("message sent");
-///
-/// // Observing the server behavior,
-/// // note that the responses come back associated to the id of the requester.
-/// assert_eq!(client_box.recv().await, Some((1, Update{from:0,to:4})));
-/// assert_eq!(client_box.recv().await, Some((2, Update{from:4,to:40})));
-/// assert_eq!(client_box.recv().await, Some((1, Update{from:40,to:42})));
-///
-/// # }
-/// ```
-///
-
-#[async_trait]
-pub trait Server: 'static + Sized + Send + Sync {
-    type Request: Message;
-    type Response: Message;
-
-    /// Return the server name
-    fn name(&self) -> &str;
-
-    /// Handle the request returning the response when done
-    ///
-    /// For such a server to return errors, the response type must be a `Result`.
-    async fn handle(&mut self, request: Self::Request) -> Self::Response;
-}
-
-#[async_trait]
-impl<S: Server> Actor for ServerActor<S> {
-    fn name(&self) -> &str {
-        self.server.name()
-    }
-
-    async fn run(&mut self) -> Result<(), RuntimeError> {
-        let server = &mut self.server;
-        while let Some((client_id, request)) = self.messages.recv().await {
-            let result = server.handle(request).await;
-            self.messages.send((client_id, result)).await?
-        }
-        Ok(())
-    }
-}
-
-/// An actor that wraps a request-response protocol
-///
-/// Requests are processed concurrently (up to some max concurrency level).
-///
-/// The server must be `Clone` to create a fresh server handle for each request.
-pub struct ConcurrentServerActor<S: Server + Clone> {
-    server: S,
-    messages: ConcurrentServerMessageBox<S::Request, S::Response>,
-}
-
-impl<S: Server + Clone> ConcurrentServerActor<S> {
-    pub fn new(server: S, messages: ConcurrentServerMessageBox<S::Request, S::Response>) -> Self {
-        ConcurrentServerActor { server, messages }
-    }
-}
-
-#[async_trait]
-impl<S: Server + Clone> Actor for ConcurrentServerActor<S> {
-    fn name(&self) -> &str {
-        self.server.name()
-    }
-
-    async fn run(&mut self) -> Result<(), RuntimeError> {
-        while let Some((client_id, request)) = self.messages.recv().await {
-            // Spawn the request
-            let mut server = self.server.clone();
-            let pending_result = tokio::spawn(async move {
-                let result = server.handle(request).await;
-                (client_id, result)
-            });
-
-            // Send the response back to the client
-            self.messages.send_response_once_done(pending_result)
-        }
-
-        Ok(())
-    }
 }
 
 #[cfg(test)]
