+++ crates/core/tedge_actors/src/lib.rs
@@ -262,22 +262,22 @@ pub mod builders;
 pub mod channels;
 pub mod converter;
 mod errors;
-mod keyed_messages;
 pub mod message_boxes;
 mod messages;
 #[doc(hidden)]
 mod run_actor;
 pub mod runtime;
+pub mod servers;
 
 pub use actors::*;
 pub use builders::*;
 pub use channels::*;
 pub use converter::*;
 pub use errors::*;
-pub use keyed_messages::*;
 pub use message_boxes::*;
 pub use messages::*;
 pub use runtime::*;
+pub use servers::*;
 
 pub use futures;
 use futures::channel::mpsc;
